file_analyzer_base/requirements.txt
numpy>=1.21.0
jieba>=0.42.1

pymupdf>=1.23.0
pdfplumber>=0.10.0
python-docx>=0.8.11
python-pptx>=0.6.21
Pillow>=9.0.0
pydub>=0.25.1

sentence-transformers>=2.2.0
torch>=1.9.0

pytesseract>=0.3.10
easyocr>=1.7.0
openai-whisper>=20231117

scikit-learn>=1.0.0
